Refactor NVIDIA GPU section for precise clock control and desktop fan support
- Implemented locked GPU and memory clock control via NVML.
- Added architecture-aware clock range detection by scanning all performance states.
- Introduced architecture-specific stepping for GPU core offsets (7.5 MHz for Ada/Blackwell, 15 MHz for others).
- Switched offset types to f32 for fractional precision support.
- Implemented per-fan manual and auto control for desktop NVIDIA GPUs.
- Updated Statistics tab to display GPU architecture, supported clock ranges, and active fan modes.
- Fixed critical multi-GPU indexing bug by using nvml_index for hardware targeting.

+++ daemon/src/dbus_interface.rs
@@ -361,12 +361,12 @@ async fn set_tdp_profile(&self, profile: &str) -> Result<(), zbus::fdo::Error> {
         }
     }
 
-    async fn set_gpu_core_offset(&self, device_index: u32, offset: i32) -> Result<(), zbus::fdo::Error> {
+    async fn set_gpu_core_offset(&self, device_index: u32, offset: f32) -> Result<(), zbus::fdo::Error> {
         crate::hardware_control::set_gpu_core_offset(device_index, offset)
             .map_err(|e| zbus::fdo::Error::Failed(e.to_string()))
     }
 
-    async fn set_gpu_memory_offset(&self, device_index: u32, offset: i32) -> Result<(), zbus::fdo::Error> {
+    async fn set_gpu_memory_offset(&self, device_index: u32, offset: f32) -> Result<(), zbus::fdo::Error> {
         crate::hardware_control::set_gpu_memory_offset(device_index, offset)
             .map_err(|e| zbus::fdo::Error::Failed(e.to_string()))
     }
